# store: Added regex-set optimization to ExpandedPostings