--- a/pkg/store/opts.go
+++ b/pkg/store/opts.go
@@ -1,3 +1,6 @@
+// Copyright 2017 The Prometheus Authors
+// Licensed under the Apache License, Version 2.0 (the "License");
+
 package store
 
 import (