<!--
    Don't forget about CHANGELOG!

    Changelog entry format:
    - [#<PR-id>](<PR-URL>) Thanos <Component> ...

    <PR-id> Id of your pull request.
    <PR-URL> URL of your PR such as https://github.com/thanos-io/thanos/pull/<PR-id>
    <Component> Component affected by your changes such as Query, Store, Receive.
-->

* [x] I added CHANGELOG entry for this change.
* [ ] Change is not relevant to the end user.

## Changes

This PR adds "regex set" optimization that's already available in Prometheus to Thanos. If label matcher is in form label=~"a|b", then we can simply get postings for label=a and label=b. This is especially useful for labels with high cardinality (as used in benchmark).

## Verification

Extended unit test and benchmark to cover this case.

```
name                                                                 old time/op    new time/op    delta
BucketIndexReader_ExpandedPostings/n="1"-4                             4.14ms ± 3%    4.15ms ± 2%     ~     (p=0.853 n=10+10)
BucketIndexReader_ExpandedPostings/n="1",j="foo"-4                     32.7ms ± 1%    33.6ms ± 1%   +2.75%  (p=0.000 n=9+8)
BucketIndexReader_ExpandedPostings/j="foo",n="1"-4                     44.6ms ±56%    34.2ms ± 2%     ~     (p=0.165 n=10+10)
BucketIndexReader_ExpandedPostings/n="1",j!="foo"-4                    24.3ms ± 2%    25.0ms ± 1%   +2.59%  (p=0.000 n=10+9)
BucketIndexReader_ExpandedPostings/i=~".*"-4                           71.1ms ± 4%    70.9ms ± 2%     ~     (p=0.971 n=10+10)
BucketIndexReader_ExpandedPostings/i=~".+"-4                            479ms ± 2%     482ms ± 2%     ~     (p=0.113 n=9+10)
BucketIndexReader_ExpandedPostings/i=~""-4                              544ms ± 0%     551ms ± 1%   +1.31%  (p=0.000 n=7+9)
BucketIndexReader_ExpandedPostings/i!=""-4                              397ms ± 2%     398ms ± 2%     ~     (p=0.529 n=10+10)
BucketIndexReader_ExpandedPostings/n="1",i=~".*",j="foo"-4             32.6ms ± 2%    33.8ms ± 2%   +3.69%  (p=0.000 n=10+10)
BucketIndexReader_ExpandedPostings/n="1",i=~".*",i!="2",j="foo"-4      41.1ms ± 2%    43.0ms ± 1%   +4.63%  (p=0.000 n=10+7)
BucketIndexReader_ExpandedPostings/n="1",i!=""-4                        183ms ± 2%     182ms ± 1%     ~     (p=0.190 n=9+9)
BucketIndexReader_ExpandedPostings/n="1",i!="",j="foo"-4                207ms ± 2%     207ms ± 2%     ~     (p=0.393 n=10+10)
BucketIndexReader_ExpandedPostings/n="1",i=~".+",j="foo"-4              289ms ± 2%     291ms ± 2%     ~     (p=0.218 n=10+10)
BucketIndexReader_ExpandedPostings/n="1",i=~"1.+",j="foo"-4            49.3ms ± 2%    48.9ms ± 2%     ~     (p=0.063 n=10+10)
BucketIndexReader_ExpandedPostings/n="1",i=~".+",i!="2",j="foo"-4       300ms ± 2%     299ms ± 2%     ~     (p=0.842 n=10+9)
BucketIndexReader_ExpandedPostings/n="1",i=~".+",i!~"2.*",j="foo"-4     334ms ± 2%     337ms ± 1%   +0.95%  (p=0.043 n=10+10)
BucketIndexReader_ExpandedPostings/i=~"0|1|2"-4                        17.8ms ± 4%     0.1ms ± 2%  -99.56%  (p=0.000 n=10+9)

name                                                                 old alloc/op   new alloc/op   delta
BucketIndexReader_ExpandedPostings/n="1"-4                             11.4MB ± 0%    11.4MB ± 0%     ~     (p=0.344 n=10+9)
BucketIndexReader_ExpandedPostings/n="1",j="foo"-4                     23.5MB ± 0%    23.5MB ± 0%     ~     (p=0.305 n=8+10)
BucketIndexReader_ExpandedPostings/j="foo",n="1"-4                     23.5MB ± 0%    23.5MB ± 0%     ~     (p=0.372 n=9+10)
BucketIndexReader_ExpandedPostings/n="1",j!="foo"-4                    23.5MB ± 0%    23.5MB ± 0%     ~     (p=0.645 n=9+9)
BucketIndexReader_ExpandedPostings/i=~".*"-4                            284MB ± 0%     284MB ± 0%     ~     (p=0.195 n=10+10)
BucketIndexReader_ExpandedPostings/i=~".+"-4                            382MB ± 0%     382MB ± 0%     ~     (p=0.081 n=10+10)
BucketIndexReader_ExpandedPostings/i=~""-4                              240MB ± 0%     240MB ± 0%   +0.00%  (p=0.002 n=9+10)
BucketIndexReader_ExpandedPostings/i!=""-4                              382MB ± 0%     382MB ± 0%     ~     (p=0.891 n=8+10)
BucketIndexReader_ExpandedPostings/n="1",i=~".*",j="foo"-4             23.5MB ± 0%    23.5MB ± 0%   +0.00%  (p=0.026 n=10+9)
BucketIndexReader_ExpandedPostings/n="1",i=~".*",i!="2",j="foo"-4      25.6MB ± 0%    25.6MB ± 0%     ~     (p=0.740 n=9+10)
BucketIndexReader_ExpandedPostings/n="1",i!=""-4                        177MB ± 0%     177MB ± 0%     ~     (all equal)
BucketIndexReader_ExpandedPostings/n="1",i!="",j="foo"-4                189MB ± 0%     189MB ± 0%     ~     (all equal)
BucketIndexReader_ExpandedPostings/n="1",i=~".+",j="foo"-4              189MB ± 0%     189MB ± 0%     ~     (p=0.083 n=9+8)
BucketIndexReader_ExpandedPostings/n="1",i=~"1.+",j="foo"-4            83.6MB ± 0%    83.6MB ± 0%   +0.00%  (p=0.014 n=10+10)
BucketIndexReader_ExpandedPostings/n="1",i=~".+",i!="2",j="foo"-4       191MB ± 0%     191MB ± 0%     ~     (p=0.698 n=9+10)
BucketIndexReader_ExpandedPostings/n="1",i=~".+",i!~"2.*",j="foo"-4     247MB ± 0%     247MB ± 0%   +0.00%  (p=0.040 n=10+10)
BucketIndexReader_ExpandedPostings/i=~"0|1|2"-4                        51.2MB ± 0%     0.0MB ± 0%  -99.97%  (p=0.000 n=10+10)

name                                                                 old allocs/op  new allocs/op  delta
BucketIndexReader_ExpandedPostings/n="1"-4                               76.0 ± 0%      76.0 ± 0%     ~     (all equal)
BucketIndexReader_ExpandedPostings/n="1",j="foo"-4                        112 ± 0%       112 ± 0%     ~     (all equal)
BucketIndexReader_ExpandedPostings/j="foo",n="1"-4                        112 ± 0%       112 ± 0%     ~     (all equal)
BucketIndexReader_ExpandedPostings/n="1",j!="foo"-4                       111 ± 0%       111 ± 0%     ~     (all equal)
BucketIndexReader_ExpandedPostings/i=~".*"-4                             96.0 ± 0%      96.0 ± 0%     ~     (all equal)
BucketIndexReader_ExpandedPostings/i=~".+"-4                             300k ± 0%      300k ± 0%   +0.00%  (p=0.000 n=9+8)
BucketIndexReader_ExpandedPostings/i=~""-4                               300k ± 0%      300k ± 0%   +0.00%  (p=0.001 n=10+10)
BucketIndexReader_ExpandedPostings/i!=""-4                               300k ± 0%      300k ± 0%     ~     (all equal)
BucketIndexReader_ExpandedPostings/n="1",i=~".*",j="foo"-4                113 ± 0%       113 ± 0%     ~     (all equal)
BucketIndexReader_ExpandedPostings/n="1",i=~".*",i!="2",j="foo"-4         145 ± 0%       145 ± 0%     ~     (all equal)
BucketIndexReader_ExpandedPostings/n="1",i!=""-4                         300k ± 0%      300k ± 0%     ~     (all equal)
BucketIndexReader_ExpandedPostings/n="1",i!="",j="foo"-4                 300k ± 0%      300k ± 0%     ~     (all equal)
BucketIndexReader_ExpandedPostings/n="1",i=~".+",j="foo"-4               300k ± 0%      300k ± 0%   +0.00%  (p=0.000 n=9+9)
BucketIndexReader_ExpandedPostings/n="1",i=~"1.+",j="foo"-4             33.5k ± 0%     33.5k ± 0%   +0.00%  (p=0.001 n=10+10)
BucketIndexReader_ExpandedPostings/n="1",i=~".+",i!="2",j="foo"-4        300k ± 0%      300k ± 0%   +0.00%  (p=0.005 n=10+9)
BucketIndexReader_ExpandedPostings/n="1",i=~".+",i!~"2.*",j="foo"-4      334k ± 0%      334k ± 0%   +0.00%  (p=0.008 n=8+10)
BucketIndexReader_ExpandedPostings/i=~"0|1|2"-4                          87.5 ± 1%     119.0 ± 0%  +36.00%  (p=0.000 n=10+10)
````