@pstibrany 2020-04-16
References:
- [Prometheus PR](https://github.com/prometheus-junkyard/tsdb/pull/602)
- [Cortex PR 1](https://github.com/cortexproject/cortex/pull/2446) and [PR 2](https://github.com/cortexproject/cortex/pull/2475)