@yeya24 2020-04-16
Nice work!
But to get the regex-set optimization, we need to upgrade Prometheus dependency to 2.17 or higher in Thanos to adapt this fix for regex match https://github.com/prometheus/prometheus/pull/6540