@pstibrany 2020-04-16
> Nice work!
> But to get the regex-set optimization, we need to upgrade Prometheus dependency to 2.17 or higher in Thanos to adapt this fix for regex match [prometheus/prometheus#6540](https://github.com/prometheus/prometheus/pull/6540)

I don't think that's true. I have modified the `findSetMatches` and its test to not require anchoring, since `matcher.Value` doesn't include that. That also means that Thanos has its own copy of `findSetMatches`, independent from Prometheus.