@pstibrany 2020-04-17
> Also we need same in Prometheus at some point.

Prometheus does this already, that's where the code was copied from.